Refactor tests in prep for Serilog additions
